// clients/typescript/src/generated/types/verificationAccountMeta.ts
/**
 * This code was AUTOGENERATED using the Codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun Codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import {
  combineCodec,
  fixDecoderSize,
  fixEncoderSize,
  getBooleanDecoder,
  getBooleanEncoder,
  getBytesDecoder,
  getBytesEncoder,
  getStructDecoder,
  getStructEncoder,
  getU8Decoder,
  getU8Encoder,
  type FixedSizeCodec,
  type FixedSizeDecoder,
  type FixedSizeEncoder,
  type ReadonlyUint8Array,
} from '@solana/kit';

export type VerificationAccountMeta = {
  discriminator: number;
  addressConfig: ReadonlyUint8Array;
  isSigner: boolean;
  isWritable: boolean;
};

export type VerificationAccountMetaArgs = VerificationAccountMeta;

export function getVerificationAccountMetaEncoder(): FixedSizeEncoder<VerificationAccountMetaArgs> {
  return getStructEncoder([
    ['discriminator', getU8Encoder()],
    ['addressConfig', fixEncoderSize(getBytesEncoder(), 32)],
    ['isSigner', getBooleanEncoder()],
    ['isWritable', getBooleanEncoder()],
  ]);
}

export function getVerificationAccountMetaDecoder(): FixedSizeDecoder<VerificationAccountMeta> {
  return getStructDecoder([
    ['discriminator', getU8Decoder()],
    ['addressConfig', fixDecoderSize(getBytesDecoder(), 32)],
    ['isSigner', getBooleanDecoder()],
    ['isWritable', getBooleanDecoder()],
  ]);
}

export function getVerificationAccountMetaCodec(): FixedSizeCodec<
  VerificationAccountMetaArgs,
  VerificationAccountMeta
> {
  return combineCodec(
    getVerificationAccountMetaEncoder(),
    getVerificationAccountMetaDecoder()
  );
}